Release 0.5.24

+++ tools/python/boutiques/tests/test_import.py
@@ -23,7 +23,8 @@ class TestImport(TestCase):
     @pytest.fixture(scope='session', autouse=True)
     def clean_up(self):
         yield
-        os.remove("user-image.simg")
+        if os.path.isfile("user-image.simg"):
+            os.remove("user-image.simg")
 
     def test_import_bids_good(self):
         bids_app = opj(op.split(bfile)[0],
@@ -59,9 +60,10 @@ class TestImport(TestCase):
         if op.isfile(fout):
             os.remove(fout)
         self.assertFalse(bosh(["import", "0.4",  fout, fin]))
-        result = open(fout, "U").read().strip()
-        self.assertIn(result, [open(ref_file, "U").read().strip(),
-                               open(ref_file_p2, "U").read().strip()])
+        result = json.loads(open(fout, "U").read().strip())
+        self.assertIn(result,
+                      [json.loads(open(ref_file, "U").read().strip()),
+                       json.loads(open(ref_file_p2, "U").read().strip())])
         os.remove(fout)
 
     def test_upgrade_04_json_obj(self):
@@ -76,9 +78,10 @@ class TestImport(TestCase):
         if op.isfile(fout):
             os.remove(fout)
         self.assertFalse(bosh(["import", "0.4",  fout, fin]))
-        result = open(fout, "U").read().strip()
-        self.assertIn(result, [open(ref_file, "U").read().strip(),
-                               open(ref_file_p2, "U").read().strip()])
+        result = json.loads(open(fout, "U").read().strip())
+        self.assertIn(result,
+                      [json.loads(open(ref_file, "U").read().strip()),
+                       json.loads(open(ref_file_p2, "U").read().strip())])
         os.remove(fout)
 
     def test_import_cwl_valid(self):
@@ -142,7 +145,7 @@ class TestImport(TestCase):
         pydocopt_input = op.join(base_path, "test_valid.py")
         descriptor_output = op.join(base_path, "test_valid_output.json")
 
-        import_args = ["import", "dcpt", descriptor_output, pydocopt_input]
+        import_args = ["import", "docopt", descriptor_output, pydocopt_input]
         bosh(import_args)
 
         test_invocation = op.join(base_path, "valid_invoc_mutex.json")
@@ -156,7 +159,7 @@ class TestImport(TestCase):
         pydocopt_input = op.join(base_path, "test_options.py")
         descriptor_output = op.join(base_path, "test_options_output.json")
 
-        import_args = ["import", "dcpt", descriptor_output, pydocopt_input]
+        import_args = ["import", "docopt", descriptor_output, pydocopt_input]
         bosh(import_args)
 
         test_invocation = op.join(base_path, "test_options_invocation.json")
@@ -170,7 +173,7 @@ class TestImport(TestCase):
         pydocopt_input = op.join(base_path, "test_invalid.py")
         descriptor_output = op.join(base_path, "foobar.json")
 
-        args = ["import", "dcpt", descriptor_output, pydocopt_input]
+        args = ["import", "docopt", descriptor_output, pydocopt_input]
 
         with pytest.raises(ImportError, match="Invalid docopt script"):
             bosh(args)
@@ -185,7 +188,7 @@ class TestImport(TestCase):
         pydocopt_input = op.join(base_path, "naval_fate.py")
         descriptor_output = op.join(base_path, "naval_fate_descriptor.json")
 
-        import_args = ["import", "dcpt", descriptor_output, pydocopt_input]
+        import_args = ["import", "docopt", descriptor_output, pydocopt_input]
         bosh(import_args)
 
         test_invocation = op.join(base_path, "nf_invoc_new.json")
